requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pdfplumber==0.11.4
Pillow==10.4.0
pytesseract==0.3.13
requests==2.32.3
python-dotenv==1.0.1
